pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nagare-clip"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "PyYAML",
    "ginza",
    "ja_ginza",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.scripts]
nagare-clip-stage2 = "nagare_clip.cli:main"

[dependency-groups]
dev = [
    "pytest",
]

[tool.hatch.build.targets.wheel]
packages = ["src/nagare_clip"]

[tool.pytest.ini_options]
testpaths = ["tests"]
